backend/pyproject.toml
[project]
name = "ai-math-tutor-backend"
version = "0.1.0"
description = "FastAPI backend for AI Math Tutor"
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.135.1",
  "uvicorn==0.41.0",
  "SQLAlchemy==2.0.48",
  "python-dotenv==1.2.2",
  "openai==2.26.0",
  "python-multipart==0.0.22",
  "pydantic==2.12.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
